% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internetdb-package.R
\docType{package}
\name{internetdb}
\alias{internetdb}
\alias{internetdb-package}
\title{Fast IP Lookups for Open Ports and Vulnerabilities}
\description{
Shodan \url{https://shodan.io/} scans the internet for hosts and services and maintains a few
APIs into that data. One API is the InternetDB \url{https://internetdb.shodan.io/}, which allows for
free querying of open ports and vulnerabilities. Tools are provided to query this servce.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/internetdb}
  \item Report bugs at \url{https://github.com/hrbrmstr/internetdb/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
