% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_internetdb}
\alias{query_internetdb}
\title{Query Shodan's InternetDB}
\usage{
query_internetdb(ip)
}
\arguments{
\item{ip}{A single IP address}
}
\value{
data frame
}
\description{
Query Shodan's InternetDB
}
\examples{
query_internetdb("36.94.222.209")
}
\references{
\link{Shodan InternetDB API Documentation}(# https://internetdb.shodan.io/docs#/)
}
